requirements.txt
# Core scraping dependencies
selenium>=4.15.0
webdriver-manager>=4.0.1
tweepy>=4.14.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# NLP and sentiment analysis
textblob>=0.17.1
nltk>=3.8.1

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0

# Optional but recommended
pillow>=10.0.0  # For image processing
aiohttp>=3.9.0  # For async requests
tqdm>=4.66.0  # Progress bars
colorama>=0.4.6  # Colored terminal output

# For proxy support
requests[socks]>=2.31.0
pysocks>=1.7.1

# Advanced analytics (optional)
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0
wordcloud>=1.9.0
